+++ SourceAFIS/Matching/EdgeConstructor.cs
@@ -11,7 +11,7 @@ namespace SourceAFIS.Matching
         {
             PolarPoint polar = Angle.ToPolar(Calc.Difference(template.Minutiae[neighbor].Position, template.Minutiae[reference].Position));
             EdgeShape edge;
-            edge.Length = (int)polar.Distance;
+            edge.Length = polar.Distance;
             edge.ReferenceAngle = Angle.Difference(template.Minutiae[reference].Direction, polar.Angle);
             edge.NeighborAngle = Angle.Difference(template.Minutiae[neighbor].Direction, Angle.Opposite(polar.Angle));
             return edge;
